feat: TakeDamage 인자 Int로 변경, OutOfScreen 매개변수 자료형 변경 및 POINT 오버로딩 삭제, PI 매크로 float형으로 변경,  UFOBeam 로직 추가

+++ CupHead/Actor.h
@@ -15,7 +15,7 @@ public:
 	virtual void Render(HDC hdc) override;				
 
 	virtual void UpdateFrame();
-	virtual void TakeDamage(float damage = 0.f) override;
+	virtual void TakeDamage(int damage = 0) override;
 	
 public:
 	inline void SetImage(Image* _Image) { image = _Image; }
