+++ CupHead/SagittariusStar.h
@@ -21,7 +21,7 @@ public:
 	virtual void Render(HDC hdc) override;
 
 	virtual void Move() override;
-	virtual void TakeDamage(float damage = 0.f) override;
+	virtual void TakeDamage(int damage = 0) override;
 
 private:
 	void Fire();
